Break out Tornado portions of TChannelRequestHandler
