+++ python/tchannel/handler.py
@@ -19,16 +19,8 @@
 # THE SOFTWARE.
 from __future__ import absolute_import
 
-import collections
-
-import tornado
-
-from .messages import CallResponseMessage
 from .messages import Types
 from .messages import PingResponseMessage
-from .messages.error import ErrorCode
-
-Endpoint = collections.namedtuple('Endpoint', ['handler', 'opts'])
 
 
 class RequestHandler(object):
@@ -101,135 +93,3 @@ class BaseRequestHandler(RequestHandler):
             implementation-specific connection object.
         """
         raise NotImplementedError("Must be implemented.")
-
-
-class TChannelRequestHandler(BaseRequestHandler):
-    """A RequestHandler that dispatches to different endpoints.
-
-    Endpoints are registered using ``register`` or the ``route``
-    decorator.
-
-    .. code-block:: python
-
-        handler = # ...
-
-        @hanlder.route('myMethod')
-        def myMethod(request, response, opts):
-            response.write('hello world')
-    """
-
-    def __init__(self):
-        super(TChannelRequestHandler, self).__init__()
-        self.endpoints = {}
-
-    def handle_call(self, message_id, call, connection):
-        request = Request(message_id, call, connection)
-        endpoint = self.endpoints.get(request.endpoint, None)
-
-        if endpoint is None:
-            return connection.send_error(
-                ErrorCode.bad_request,
-                "Endpoint '%s' for service '%s' is not defined" % (
-                    call.args[0], call.service
-                ),
-                message_id,
-            )
-
-        response = Response(message_id, connection)
-        result = None
-        try:
-            result = endpoint.handler(request, response, endpoint.opts)
-            if isinstance(result, tornado.gen.Future):
-                result.add_done_callback(lambda f: response.finish())
-                tornado.ioloop.IOLoop.current().add_future(
-                    result,
-                    lambda f: f.exception()
-                )
-            return result
-        # TODO add tchannel error handling here
-        finally:
-            if result is None:
-                response.finish()
-
-    def route(self, rule, **opts):
-        def decorator(handler):
-            self.register(rule, handler, **opts)
-            return handler
-
-        return decorator
-
-    def register(self, rule, handler, **opts):
-        """Register a new endpoint with the given name.
-
-        .. code-block:: python
-
-            def handler(request, response, opts):
-                print opts  # => {'foo': 'bar'}
-                # ...
-
-            handler.register('is_healthy', handler, foo='bar')
-
-        :param rule:
-            Name of the endpoint. Incoming Call Requests must have this as
-            ``arg1`` to dispatch to this handler.
-        :param handler:
-            A function that gets called with ``Request``, ``Response``, and
-            the ``opts`` dictionary.
-        :param opts:
-            Parameters to pass to the ``handler`` as a dictionary.
-        """
-        self.endpoints[rule] = Endpoint(handler=handler, opts=opts)
-
-
-class Request(object):
-    """Represents an incoming request to an endpoint."""
-
-    __slots__ = ('message', 'header', 'body', 'endpoint', 'id', 'connection')
-
-    def __init__(self, message_id, message, connection):
-        assert len(message.args) == 3
-
-        self.id = message_id
-        self.message = message
-        self.connection = connection
-
-        self.endpoint = self.message.args[0]
-        self.header = self.message.args[1]
-        self.body = self.message.args[2]
-
-
-class Response(object):
-    """An outgoing response.
-
-    Handlers will set either ``write`` or manually set the ``message`` to
-    specify the response message.
-    """
-
-    __slots__ = ('connection', 'message', 'id', 'arg1', 'arg2', 'arg3')
-
-    def __init__(self, message_id, connection):
-        self.id = message_id
-        self.connection = connection
-
-        self.arg1 = ""
-        self.arg2 = ""
-        self.arg3 = ""
-
-        self.message = None
-
-    def write(self, arg1="", arg2="", arg3=""):
-        """Write the given args to the response."""
-        self.arg1 += arg1
-        self.arg2 += arg2
-        self.arg3 += arg3
-
-    def finish(self):
-        """Finish writing the response."""
-        # TODO failure codes
-
-        if self.message is None:
-            self.message = CallResponseMessage(
-                args=[self.arg1, self.arg2, self.arg3]
-            )
-
-        return self.connection.frame_and_write_stream(self.message, self.id)
